+++ blocksuite/affine/gfx/turbo-renderer/src/index.ts
@@ -1,4 +1,3 @@
-export * from './extension';
 export * from './layout/block-layout-provider';
 export * from './painter/painter.worker';
 export * from './text-utils';
